test(x86len): arm the guard page only around each decode
x86len_test installed its SIGBUS/SIGSEGV handlers for the whole run,
so a fault anywhere outside mx_decode would have siglongjmp'd into a
frame that had already returned. The handlers are now installed just
before each guarded decode and the previous ones restored just after.
A new check that both are back to SIG_DFL once the decodes are done
fails without the restore; with the decoder mutated to read past
`avail`, the guard still reports "read past avail".

x86len_oracle sized its function-starts array as (datasize + 1) in 32
bits, which wraps at a datasize of UINT32_MAX; the sum is now taken in
size_t. Nothing can falsify this short of a 4 GiB image: the payload
is already bounded by the file's size.

+++ tests/x86len_oracle.c
@@ -168,7 +168,7 @@ int main(int argc, char **argv) {
 
     uint64_t lo = f.text->addr, hi = f.text->addr + f.text->size;
     const uint8_t *code = im.buf + f.text->offset;
-    uint64_t *starts = (uint64_t *)malloc((f.fs->datasize + 1) * sizeof *starts);
+    uint64_t *starts = (uint64_t *)malloc(((size_t)f.fs->datasize + 1) * sizeof *starts);
     if (!starts) { fprintf(stderr, "x86len_oracle: out of memory\n"); return 2; }
     size_t nstarts = 0;
     const uint8_t *p = im.buf + f.fs->dataoff, *pe = p + f.fs->datasize;
